// src/app/models/MetroArea.ts
export class MetroArea {
  id: number;
  name: string;
  hourly_pay_stat_coverage_percent?: number;
  hourly_pay_stat_coverage_percent_delivery?: number;
  hourly_pay_stat_coverage_percent_rideshare?: number;
  hourly_pay_stat_coverage_count_workers?: number;
  hourly_pay_stat_coverage_count_workers_rideshare?: number;
  hourly_pay_stat_coverage_count_workers_delivery?: number;
  hourly_pay_stat_coverage_count_employers?: number;
  hourly_pay_stat_coverage_count_employers_rideshare?: number;
  hourly_pay_stat_coverage_count_employers_delivery?: number;
  hourly_pay_stat_coverage_count_jobs?: number;
  hourly_pay_stat_coverage_count_jobs_rideshare?: number;
  hourly_pay_stat_coverage_count_jobs_delivery?: number;
  hourly_pay_stat_coverage_count_tasks?: number;
  hourly_pay_stat_coverage_count_tasks_rideshare?: number;
  hourly_pay_stat_coverage_count_tasks_delivery?: number;
}